+++ test/integration/linux/ssh_session_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/veteranbv/hostinfo/internal/collectors"
+	"github.com/veteranbv/sysgreet/internal/collectors"
 )
 
 func TestSessionCollectorUsesSSHConnection(t *testing.T) {